Create Pairs of Songs with total durations divisible by 60
This problem asks us to count pairs of songs whose total duration is divisible by 60. We are given an array time, where each element represents the duration of a song. We need to find the number of valid pairs (i, j) such that i < j and (time[i] + time[j]) % 60 == 0. The key observation is that only the remainder of each song length modulo 60 matters. For each song, its complement remainder is (60 - remainder) % 60. We can use a frequency array of size 60 to store counts of remainders and efficiently compute valid pairs. This ensures an O(n) solution.