Valid Anagram problem 242
 C solution for the "Valid Anagram" problem.

Here is the code in a new file. It uses a character frequency array for a highly efficient solution that will pass all test cases.

I've created a new file with the C solution. This code first checks if the string lengths are equal. Then, it uses a single array to track the character counts, incrementing for the first string and decrementing for the second. If the strings are anagrams, all counts in the array will be zero at the end. This is a very fast and memory-efficient approach.